+++ src/api/gateway.go
@@ -98,13 +98,13 @@ type Visorer interface {
 // Walleter interface for wallet.Service methods used by the API
 type Walleter interface {
 	UnloadWallet(wltID string) error
-	EncryptWallet(wltID string, password []byte) (*wallet.Wallet, error)
-	DecryptWallet(wltID string, password []byte) (*wallet.Wallet, error)
+	EncryptWallet(wltID string, password []byte) (wallet.Wallet, error)
+	DecryptWallet(wltID string, password []byte) (wallet.Wallet, error)
 	GetWalletSeed(wltID string, password []byte) (string, error)
-	CreateWallet(wltName string, options wallet.Options, bg wallet.TransactionsFinder) (*wallet.Wallet, error)
-	RecoverWallet(wltID, seed string, password []byte) (*wallet.Wallet, error)
+	CreateWallet(wltName string, options wallet.Options, bg wallet.TransactionsFinder) (wallet.Wallet, error)
+	RecoverWallet(wltID, seed string, password []byte) (wallet.Wallet, error)
 	NewAddresses(wltID string, password []byte, n uint64) ([]cipher.Address, error)
-	GetWallet(wltID string) (*wallet.Wallet, error)
+	GetWallet(wltID string) (wallet.Wallet, error)
 	GetWallets() (wallet.Wallets, error)
 	UpdateWalletLabel(wltID, label string) error
 	WalletDir() (string, error)