+++ src/gui/static/src/app/app.module.ts
@@ -111,6 +111,7 @@ import { ChangeNoteComponent } from './components/pages/send-skycoin/send-previe
 import { MsgBarComponent } from './components/layout/msg-bar/msg-bar.component';
 import { MsgBarService } from './services/msg-bar.service';
 import { SeedWordDialogComponent } from './components/layout/seed-word-dialog/seed-word-dialog.component';
+import { MultipleDestinationsDialogComponent } from './components/layout/multiple-destinations-dialog/multiple-destinations-dialog.component';
 
 
 const ROUTES = [
@@ -250,6 +251,7 @@ const ROUTES = [
     ChangeNoteComponent,
     MsgBarComponent,
     SeedWordDialogComponent,
+    MultipleDestinationsDialogComponent,
   ],
   entryComponents: [
     AddDepositAddressComponent,
@@ -282,6 +284,7 @@ const ROUTES = [
     HwUpdateAlertDialogComponent,
     ChangeNoteComponent,
     SeedWordDialogComponent,
+    MultipleDestinationsDialogComponent,
   ],
   imports: [
     BrowserModule,