+++ src/wallet/readable.go
@@ -1,11 +1,7 @@
 package wallet
 
 import (
-	"fmt"
-	"strconv"
-
 	"github.com/skycoin/skycoin/src/cipher"
-	"github.com/skycoin/skycoin/src/util/file"
 )
 
 // ReadableEntry wallet entry with json tags
@@ -43,7 +39,20 @@ func NewReadableEntry(coinType CoinType, w Entry) ReadableEntry {
 // ReadableEntries array of ReadableEntry
 type ReadableEntries []ReadableEntry
 
-// ToWalletEntries convert readable entries to entries
+func newReadableEntries(entries Entries, coinType CoinType) ReadableEntries {
+	re := make(ReadableEntries, len(entries))
+	for i, e := range entries {
+		re[i] = NewReadableEntry(coinType, e)
+	}
+	return re
+}
+
+// GetEntries returns this array
+func (res ReadableEntries) GetEntries() ReadableEntries {
+	return res
+}
+
+// toWalletEntries convert readable entries to entries
 // converts base on the wallet version.
 func (res ReadableEntries) toWalletEntries(coinType CoinType, isEncrypted bool) ([]Entry, error) {
 	entries := make([]Entry, len(res))
@@ -111,86 +120,12 @@ func newEntryFromReadable(coinType CoinType, w *ReadableEntry) (*Entry, error) {
 	}, nil
 }
 
-// ReadableWallet used for [de]serialization of a Wallet
-type ReadableWallet struct {
-	Meta    map[string]string `json:"meta"`
-	Entries ReadableEntries   `json:"entries"`
-}
-
-// NewReadableWallet creates readable wallet
-func NewReadableWallet(w *Wallet) *ReadableWallet {
-	readable := make(ReadableEntries, len(w.Entries))
-	for i, e := range w.Entries {
-		readable[i] = NewReadableEntry(w.coin(), e)
-	}
-
-	meta := make(map[string]string, len(w.Meta))
-	for k, v := range w.Meta {
-		meta[k] = v
-	}
-
-	return &ReadableWallet{
-		Meta:    meta,
-		Entries: readable,
-	}
-}
-
-// LoadReadableWallet loads a ReadableWallet from disk
-func LoadReadableWallet(filename string) (*ReadableWallet, error) {
-	w := &ReadableWallet{}
-	if err := w.Load(filename); err != nil {
-		return nil, fmt.Errorf("load wallet %s failed: %v", filename, err)
-	}
-	return w, nil
-}
-
-// ToWallet convert readable wallet to Wallet
-func (rw *ReadableWallet) ToWallet() (*Wallet, error) {
-	w := &Wallet{
-		Meta: rw.Meta,
-	}
-
-	if err := w.Validate(); err != nil {
-		return nil, fmt.Errorf("invalid wallet %s: %v", w.Filename(), err)
-	}
-
-	ets, err := rw.Entries.toWalletEntries(w.coin(), w.IsEncrypted())
-	if err != nil {
-		return nil, err
-	}
-
-	w.Entries = ets
-
-	return w, nil
-}
-
-// Save saves to filename
-func (rw *ReadableWallet) Save(filename string) error {
-	return file.SaveJSON(filename, rw, 0600)
-}
-
-// Load loads from filename
-func (rw *ReadableWallet) Load(filename string) error {
-	return file.LoadJSON(filename, rw)
-}
-
-func (rw *ReadableWallet) timestamp() int64 {
-	// Intentionally ignore the error when parsing the timestamp,
-	// if it isn't valid or is missing it will be set to 0
-	x, _ := strconv.ParseInt(rw.Meta[metaTimestamp], 10, 64) //nolint:errcheck
-	return x
-}
-
-func (rw *ReadableWallet) filename() string {
-	return rw.Meta[metaFilename]
-}
-
-// Erase remove sensitive data
-func (rw *ReadableWallet) Erase() {
-	delete(rw.Meta, metaSeed)
-	delete(rw.Meta, metaLastSeed)
-	delete(rw.Meta, metaSecrets)
-	for i := range rw.Entries {
-		rw.Entries[i].Secret = ""
-	}
+// Readable defines the readable wallet API.
+// A readable wallet is the on-disk representation of a wallet.
+type Readable interface {
+	ToWallet() (Wallet, error)
+	Timestamp() int64
+	SetFilename(string)
+	Filename() string
+	GetEntries() ReadableEntries
 }